\name{guano-package}
\alias{guano-package}
\alias{guano}
\docType{package}
\title{
\packageTitle{guano}
}
\description{
\packageDescription{guano}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{guano}
\packageIndices{guano}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{guano}

Maintainer: \packageMaintainer{guano}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
