% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guano.R
\name{read.guano.dir}
\alias{read.guano.dir}
\title{Read all GUANO file in a directory as a single dataframe}
\usage{
read.guano.dir(dirname, pattern = "*.wav", recursive = FALSE)
}
\arguments{
\item{dirname}{The directory name (or path) which contains GUANO files}

\item{pattern}{An optional glob pattern. Only files names which match will be returned (default "*.wav")}

\item{recursive}{logical. Should we recurse into sub-directories?}
}
\value{
dataframe with metadata attributes as columns
}
\description{
Read all GUANO file in a directory as a single dataframe
}
