% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guano.R
\name{read.guano}
\alias{read.guano}
\title{Read a single GUANO file}
\usage{
read.guano(filename)
}
\arguments{
\item{filename}{The GUANO filename or path}
}
\value{
list of named metadata fields
}
\description{
Read a single GUANO file
}
